# docs/swagger.yaml
definitions:
  fitness-tracker-backend_user_handler.errorResponse:
    properties:
      error:
        type: string
    type: object
...
  handler.createUserRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    required:
    - email
    - name
    - password
    type: object
  handler.loginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
  handler.refreshRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
...
  handler.tokenResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
...
  handler.updateUserRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
...
  models.User:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      passwordHash:
        type: string
    type: object
...
info:
  contact: {}
  description: API documentation for the Fitness Tracker backend service.
  title: Fitness Tracker API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates user credentials and returns JWT pair
      parameters:
      - description: Credentials
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handler.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.tokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fitness-tracker-backend_user_handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/fitness-tracker-backend_user_handler.errorResponse'
      summary: User login
      tags:
      - auth
  /auth/logout:
    post:
      responses:
        "204":
          description: No Content
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Logout (client-side token discard)
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Exchanges a refresh token for a new JWT pair
      parameters:
      - description: Refresh token
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handler.refreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.tokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fitness-tracker-backend_user_handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/fitness-tracker-backend_user_handler.errorResponse'
      summary: Refresh JWT tokens
      tags:
      - auth
...
  /users:
    get:
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/fitness-tracker-backend_user_handler.errorResponse'
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Creates a user and returns the stored record
      parameters:
      - description: User info
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handler.createUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fitness-tracker-backend_user_handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/fitness-tracker-backend_user_handler.errorResponse'
      summary: Register new user
      tags:
      - users
  /users/{id}:
    delete:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fitness-tracker-backend_user_handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/fitness-tracker-backend_user_handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - users
    get:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fitness-tracker-backend_user_handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/fitness-tracker-backend_user_handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update information
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handler.updateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fitness-tracker-backend_user_handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/fitness-tracker-backend_user_handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/fitness-tracker-backend_user_handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - users
  /users/me:
    get:
      description: Returns the authenticated user's information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/fitness-tracker-backend_user_handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/fitness-tracker-backend_user_handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - users
...
securityDefinitions:
  BearerAuth:
    description: 'Provide your JWT with the "Bearer " prefix. Example: "Bearer {token}".'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"